refactor : 리뷰 작성 api 코드 정리
