--- dto/review/UpdateReviewRequest.js
@@ -1,9 +1,0 @@
-class UpdateReviewRequest {
-    constructor({rating, keywords, review}) {
-        this.rating = rating;
-        this.keywords = keywords;
-        this.review = review;
-    }
-}
-
-module.exports = UpdateReviewRequest;
